Split NFA into smaller source files
