+++ src/nfa/nfa.cc
@@ -43,99 +43,6 @@ const Symbol Limits::min_symbol;
 const Symbol Limits::max_symbol;
 
 namespace {
-    Simlib::Util::BinaryRelation compute_fw_direct_simulation(const Nfa& aut) {
-        Symbol maxSymbol = 0;
-        const size_t state_num = aut.size();
-        Simlib::ExplicitLTS LTSforSimulation(state_num);
-
-        for (const auto& transition : aut.delta) {
-            LTSforSimulation.add_transition(transition.src, transition.symb, transition.tgt);
-            if (transition.symb > maxSymbol) {
-                maxSymbol = transition.symb;
-            }
-        }
-
-        // final states cannot be simulated by nonfinal -> we add new selfloops over final states with new symbol in LTS
-        for (State finalState : aut.final) {
-            LTSforSimulation.add_transition(finalState, maxSymbol + 1, finalState);
-        }
-
-        LTSforSimulation.init();
-        return LTSforSimulation.compute_simulation();
-    }
-
-	Nfa reduce_size_by_simulation(const Nfa& aut, StateToStateMap &state_map) {
-        Nfa result;
-        const auto sim_relation = Algorithms::compute_relation(
-                aut, StringMap{{"relation", "simulation"}, {"direction", "forward"}});
-
-        auto sim_relation_symmetric = sim_relation;
-        sim_relation_symmetric.restrict_to_symmetric();
-
-        // for State q, quot_proj[q] should be the representative state representing the symmetric class of states in simulation
-        std::vector<size_t> quot_proj;
-        sim_relation_symmetric.get_quotient_projection(quot_proj);
-
-		const size_t num_of_states = aut.size();
-
-		// map each state q of aut to the state of the reduced automaton representing the simulation class of q
-		for (State q = 0; q < num_of_states; ++q) {
-			const State qReprState = quot_proj[q];
-			if (state_map.count(qReprState) == 0) { // we need to map q's class to a new state in reducedAut
-				const State qClass = result.add_state();
-				state_map[qReprState] = qClass;
-				state_map[q] = qClass;
-			} else {
-				state_map[q] = state_map[qReprState];
-			}
-		}
-
-        for (State q = 0; q < num_of_states; ++q) {
-            const State q_class_state = state_map.at(q);
-
-            if (aut.initial[q]) { // if a symmetric class contains initial state, then the whole class should be initial
-                result.initial.insert(q_class_state);
-            }
-
-            if (quot_proj[q] == q) { // we process only transitions starting from the representative state, this is enough for simulation
-                for (const auto &q_trans : aut.get_moves_from(q)) {
-                    const StateSet representatives_of_states_to = [&]{
-                        StateSet state_set;
-                        for (auto s : q_trans.targets) {
-                            state_set.insert(quot_proj[s]);
-                        }
-                        return state_set;
-                    }();
-
-                    // get the class states of those representatives that are not simulated by another representative in representatives_of_states_to
-                    StateSet representatives_class_states;
-                    for (const State s : representatives_of_states_to) {
-                        bool is_state_important = true; // if true, we need to keep the transition from q to s
-                        for (const State p : representatives_of_states_to) {
-                            if (s != p && sim_relation.get(s, p)) { // if p (different from s) simulates s
-                                is_state_important = false; // as p simulates s, the transition from q to s is not important to keep, as it is subsumed in transition from q to p
-                                break;
-                            }
-                        }
-                        if (is_state_important) {
-                            representatives_class_states.insert(state_map.at(s));
-                        }
-                    }
-
-                    // add the transition 'q_class_state-q_trans.symbol->representatives_class_states' at the end of transition list of transitions starting from q_class_state
-                    // as the q_trans.symbol should be the largest symbol we saw (as we iterate trough getTransitionsFromState(q) which is ordered)
-                    result.delta.get_mutable_post(q_class_state).insert(Move(q_trans.symbol, representatives_class_states));
-                }
-
-                if (aut.final[q]) { // if q is final, then all states in its class are final => we make q_class_state final
-                    result.final.insert(q_class_state);
-                }
-            }
-        }
-
-        return result;
-    }
-
     /**
      * Compute reachability of states.
      *
@@ -287,294 +194,6 @@ namespace {
         }
     }
 }
-
-std::ostream &std::operator<<(std::ostream &os, const Mata::Nfa::Trans &trans) { // {{{
-    std::string result = "(" + std::to_string(trans.src) + ", " +
-                         std::to_string(trans.symb) + ", " + std::to_string(trans.tgt) + ")";
-    return os << result;
-}
-
-size_t Delta::size() const
-{
-    size_t size = 0;
-    for (State q = 0; q < num_of_states(); ++q)
-        for (const Move & m: (*this)[q])
-            size = size + m.size();
-
-    return size;
-}
-
-void Delta::add(State state_from, Symbol symbol, State state_to) {
-    const State max_state{ std::max(state_from, state_to) };
-    if (max_state >= posts.size()) {
-        reserve_on_insert(posts, max_state);
-        posts.resize(max_state + 1);
-    }
-
-    Post& post_to_add_to{ posts[state_from] };
-
-    if (post_to_add_to.empty()) {
-        post_to_add_to.insert({ symbol, state_to });
-    } else if (post_to_add_to.back().symbol < symbol) {
-        post_to_add_to.insert({ symbol, state_to });
-    } else {
-        const auto move_it{ post_to_add_to.find(symbol) };
-        if (move_it != post_to_add_to.end()) {
-            // Add transition with symbol already used on transitions from state_from.
-            move_it->insert(state_to);
-        } else {
-            // Add transition to a new Move struct with symbol yet unused on transitions from state_from.
-            const Move new_move{ symbol, state_to };
-            post_to_add_to.insert(new_move);
-        }
-    }
-}
-
-void Delta::add(const State state_from, const Symbol symbol, const StateSet& states) {
-    if(states.empty()) {
-        return;
-    }
-
-    const State max_state{ std::max(state_from, states.back()) };
-    if (max_state >= posts.size()) {
-        reserve_on_insert(posts, max_state + 1);
-        posts.resize(max_state + 1);
-    }
-
-    Post& state_from_post{ posts[state_from] };
-
-    if (state_from_post.empty()) {
-        state_from_post.insert({ symbol, states });
-    } else if (state_from_post.back().symbol < symbol) {
-        state_from_post.insert({ symbol, states });
-    } else {
-        const auto move_it{ state_from_post.find(symbol) };
-        if (move_it != state_from_post.end()) {
-            // Add transition with symbolOnTransition already used on transitions from state_from.
-            move_it->insert(states);
-
-        } else {
-            // Add transition to a new Move struct with symbol yet unused on transitions from state_from.
-            // Move new_symbol_transitions{ symbol, states };
-            state_from_post.insert(Move{ symbol, states});
-        }
-    }
-}
-
-void Delta::remove(State src, Symbol symb, State tgt) {
-    if (src >= posts.size()) {
-        return;
-    }
-
-    Post& post_to_remove_from{ posts[src] };
-    if (post_to_remove_from.empty()) {
-        throw std::invalid_argument(
-                "Transition [" + std::to_string(src) + ", " + std::to_string(symb) + ", " +
-                std::to_string(tgt) + "] does not exist.");
-    } else if (post_to_remove_from.back().symbol < symb) {
-        throw std::invalid_argument(
-                "Transition [" + std::to_string(src) + ", " + std::to_string(symb) + ", " +
-                std::to_string(tgt) + "] does not exist.");
-    } else {
-        const auto move_it{ post_to_remove_from.find(symb) };
-        if (move_it == post_to_remove_from.end()) {
-            throw std::invalid_argument(
-                    "Transition [" + std::to_string(src) + ", " + std::to_string(symb) + ", " +
-                    std::to_string(tgt) + "] does not exist.");
-        } else {
-            move_it->remove(tgt);
-            if (move_it->empty()) {
-                posts[src].remove(*move_it);
-            }
-        }
-    }
-}
-
-bool Delta::contains(State src, Symbol symb, State tgt) const
-{ // {{{
-    if (posts.empty()) {
-        return false;
-    }
-
-    if (posts.size() <= src)
-        return false;
-
-    const Post& tl = posts[src];
-    if (tl.empty()) {
-        return false;
-    }
-    auto move_it{ tl.find(symb) };
-    if (move_it == tl.cend()) {
-        return false;
-    }
-
-    return move_it->targets.find(tgt) != move_it->targets.end();
-}
-
-bool Delta::empty() const
-{
-    return this->begin() == this->end();
-}
-
-Delta::const_iterator::const_iterator(const std::vector<Post>& post_p, bool ise) :
-    post(post_p), current_state(0), is_end{ ise }
-{
-    const size_t post_size = post.size();
-    for (size_t i = 0; i < post_size; ++i) {
-        if (!post[i].empty()) {
-            current_state = i;
-            post_iterator = post[i].begin();
-            targets_position = post_iterator->targets.begin();
-            return;
-        }
-    }
-
-    // no transition found, an empty post
-    is_end = true;
-}
-
-Delta::const_iterator& Delta::const_iterator::operator++()
-{
-    assert(post.begin() != post.end());
-
-    ++targets_position;
-    if (targets_position != post_iterator->targets.end())
-        return *this;
-
-    ++post_iterator;
-    if (post_iterator != post[current_state].cend()) {
-        targets_position = post_iterator->targets.begin();
-        return *this;
-    }
-
-    ++current_state;
-    while (current_state < post.size() && post[current_state].empty()) // skip empty posts
-        current_state++;
-
-    if (current_state >= post.size())
-        is_end = true;
-    else {
-        post_iterator = post[current_state].begin();
-        targets_position = post_iterator->targets.begin();
-    }
-
-    return *this;
-}
-
-Delta::const_iterator Delta::const_iterator::operator++(int) {
-    const const_iterator tmp = *this;
-    ++(*this);
-    return tmp;
-}
-
-Delta::const_iterator& Delta::const_iterator::operator=(const Delta::const_iterator& x) {
-    this->post_iterator = x.post_iterator;
-    this->targets_position = x.targets_position;
-    this->current_state = x.current_state;
-    this->is_end = x.is_end;
-
-    return *this;
-}
-
-bool Mata::Nfa::operator==(const Delta::const_iterator& a, const Delta::const_iterator& b) {
-    if (a.is_end && b.is_end) {
-        return true;
-    } else if ((a.is_end && !b.is_end) || (!a.is_end && b.is_end)) {
-        return false;
-    } else {
-        return a.current_state == b.current_state && a.post_iterator == b.post_iterator
-               && a.targets_position == b.targets_position;
-    }
-}
-
-State Delta::find_max_state() {
-    size_t max = 0;
-    State src = 0;
-    for (Post & p: posts) {
-        if (src > max)
-            max = src;
-        for (Move & m: p) {
-            if (!m.targets.empty())
-                if (m.targets.back() > max)
-                    max = m.targets.back();
-        }
-        src++;
-    }
-    return max;
-}
-
-std::vector<Post> Delta::transform(const std::function<State(State)>& lambda) const {
-    std::vector<Post> cp_post_vector;
-    cp_post_vector.reserve(num_of_states());
-    for(const Post& act_post: this->posts) {
-        Post cp_post;
-        cp_post.reserve(act_post.size());
-        for(const Move& mv : act_post) {
-            StateSet cp_dest;
-            cp_dest.reserve(mv.size());
-            for(const State& state : mv.targets) {
-                cp_dest.push_back(std::move(lambda(state)));
-            }
-            cp_post.push_back(std::move(Move(mv.symbol, cp_dest)));
-        }
-        cp_post_vector.emplace_back(cp_post);
-    }
-    return cp_post_vector;
-}
-
-Post& Delta::get_mutable_post(State q) {
-    if (q >= posts.size()) {
-        Util::reserve_on_insert(posts, q);
-        const size_t new_size{ q + 1 };
-        posts.resize(new_size);
-    }
-
-    return posts[q];
-}
-
-void Delta::defragment(const BoolVector& is_staying, const std::vector<State>& renaming) {
-    //TODO: this function seems to be unreadable, should be refactored, maybe into several functions with a clear functionality?
-
-    //first, indexes of post are filtered (places of to be removed states are taken by states on their right)
-    size_t move_index{ 0 };
-    std::erase_if(posts,
-         [&](Post&) -> bool {
-             size_t prev{ move_index };
-             ++move_index;
-             return !is_staying[prev];
-         }
-    );
-
-    //this iterates through every post and every move, filters and renames states,
-    //and then removes moves that became empty.
-    for (State q=0,size=posts.size(); q < size; ++q) {
-        Post & p = get_mutable_post(q);
-        for (auto move = p.begin(); move < p.end(); ++move) {
-            move->targets.erase(
-                    std::remove_if(move->targets.begin(), move->targets.end(), [&](State q) -> bool {
-                        return !is_staying[q];
-                    }),
-                    move->targets.end()
-            );
-            move->targets.rename(renaming);
-        }
-        p.erase(
-                std::remove_if(p.begin(), p.end(), [&](Move& move) -> bool {
-                    return move.targets.empty();
-                }),
-                p.end()
-        );
-    }
-}
-
-const Post& Delta::operator[](State q) const {
-    if (q >= posts.size()) {
-        return empty_post;
-    }
-    return posts[q];
-}
-
-///// Nfa structure related methods.
 
 void Nfa::remove_epsilon(const Symbol epsilon)
 {
@@ -777,560 +396,6 @@ StateSet Nfa::get_useful_states_old() const
     return useful_states;
 }
 
-// General methods for NFA.
-
-bool Mata::Nfa::are_state_disjoint(const Nfa& lhs, const Nfa& rhs)
-{ // {{{
-    // fill lhs_states with all states of lhs
-    std::unordered_set<State> lhs_states;
-    lhs_states.insert(lhs.initial.begin(), lhs.initial.end());
-    lhs_states.insert(lhs.final.begin(), lhs.final.end());
-
-    const size_t delta_size = lhs.delta.num_of_states();
-    for (size_t i = 0; i < delta_size; i++) {
-        lhs_states.insert(i);
-        for (const auto& symStates : lhs.delta[i])
-        {
-            lhs_states.insert(symStates.targets.begin(), symStates.targets.end());
-        }
-    }
-
-    // for every state found in rhs, check its presence in lhs_states
-    for (const auto& rhs_st : rhs.initial) {
-        if (haskey(lhs_states, rhs_st)) { return false; }
-    }
-
-    for (const auto& rhs_st : rhs.final) {
-        if (haskey(lhs_states, rhs_st)) { return false; }
-    }
-
-    const size_t lhs_post_size = lhs.delta.num_of_states();
-    for (size_t i = 0; i < lhs_post_size; i++) {
-        if (haskey(lhs_states, i))
-            return false;
-        for (const auto& symState : lhs.delta[i]) {
-            for (const auto& rhState : symState.targets) {
-                if (haskey(lhs_states,rhState)) {
-                    return false;
-                }
-            }
-        }
-    }
-
-    // no common state found
-    return true;
-} // are_disjoint }}}
-
-bool Mata::Nfa::make_complete(Nfa& aut, const Alphabet& alphabet, State sink_state) {
-    return Mata::Nfa::make_complete(aut, alphabet.get_alphabet_symbols(), sink_state);
-}
-
-bool Mata::Nfa::make_complete(Nfa& aut, const Mata::Util::OrdVector<Symbol>& symbols, State sink_state) {
-    bool was_something_added{ false };
-
-    size_t num_of_states{ aut.size() };
-    for (State state = 0; state < num_of_states; ++state) {
-        OrdVector<Symbol> used_symbols{};
-        for (auto const &move : aut.delta[state]) {
-            used_symbols.insert(move.symbol);
-        }
-        Mata::Util::OrdVector<Symbol> unused_symbols{ symbols.difference(used_symbols) };
-        for (Symbol symb : unused_symbols) {
-            aut.delta.add(state, symb, sink_state);
-            was_something_added = true;
-        }
-    }
-
-    if (was_something_added && num_of_states <= sink_state) {
-        for (Symbol symbol : symbols) {
-            aut.delta.add(sink_state, symbol, sink_state);
-        }
-    }
-
-    return was_something_added;
-}
-
-//TODO: based on the comments inside, this function needs to be rewritten in a more optimal way.
-Nfa Mata::Nfa::remove_epsilon(const Nfa& aut, Symbol epsilon) {
-    // cannot use multimap, because it can contain multiple occurrences of (a -> a), (a -> a)
-    std::unordered_map<State, StateSet> eps_closure;
-
-    // TODO: grossly inefficient
-    // First we compute the epsilon closure.
-    const size_t num_of_states{ aut.size() };
-    for (size_t i{ 0 }; i < num_of_states; ++i) {
-        for (const Move& move: aut.delta[i]) {
-            const auto it_ins_pair = eps_closure.insert({i, {i}});  // Initialize.
-            if (move.symbol == epsilon) {
-                StateSet& closure = it_ins_pair.first->second;
-                // TODO: Fix possibly insert to OrdVector. Create list already ordered, then merge (do not need to resize each time);
-                closure.insert(move.targets);
-            }
-        }
-    }
-
-    bool changed = true;
-    while (changed) { // Compute the fixpoint.
-        changed = false;
-        for (size_t i = 0; i < num_of_states; ++i) {
-            const Post& post{ aut.delta[i] };
-            const auto eps_move_it { post.find(Move{ epsilon}) };//TODO: make faster if default epsilon
-            if (eps_move_it != post.end()) {
-                StateSet& src_eps_cl = eps_closure[i];
-                for (const State tgt: eps_move_it->targets) {
-                    const StateSet& tgt_eps_cl = eps_closure[tgt];
-                    for (const State st: tgt_eps_cl) {
-                        if (src_eps_cl.count(st) == 0) {
-                            changed = true;
-                            break;
-                        }
-                    }
-                    src_eps_cl.insert(tgt_eps_cl);
-                }
-            }
-        }
-    }
-
-    // Construct the automaton without epsilon transitions.
-    Nfa result{ Delta{}, aut.initial, aut.final, aut.alphabet };
-    for (const auto& state_closure_pair : eps_closure) { // For every state.
-        State src_state = state_closure_pair.first;
-        for (State eps_cl_state : state_closure_pair.second) { // For every state in its epsilon closure.
-            if (aut.final[eps_cl_state]) result.final.insert(src_state);
-            for (const Move& move : aut.delta[eps_cl_state]) {
-                if (move.symbol == epsilon) continue;
-                // TODO: this could be done more efficiently if we had a better add method
-                for (State tgt_state : move.targets) {
-                    result.delta.add(src_state, move.symbol, tgt_state);
-                }
-            }
-        }
-    }
-    return result;
-}
-
-Nfa Mata::Nfa::fragile_revert(const Nfa& aut) {
-    const size_t num_of_states{ aut.size() };
-
-    Nfa result(num_of_states);
-
-    result.initial = aut.final;
-    result.final = aut.initial;
-
-    //COMPUTE NON-e SYMBOLS
-    //This stuff is ugly because of our handling of epsilon, what about negative numbers?
-    OrdVector<Symbol> symbols = aut.get_used_symbols();
-    if (symbols.empty())
-        return result;
-    if (symbols.back() == EPSILON)
-        symbols.pop_back();
-    // size of the "used alphabet", i.e. max symbol+1 or 0
-    Symbol alphasize =  (symbols.empty()) ? 0 : (symbols.back()+1);
-
-#ifdef _STATIC_STRUCTURES_
-    //STATIC DATA STRUCTURES:
-    // Not sure that it works ideally, whether the space for the inner vectors stays there.
-    static std::vector<std::vector<State>> sources;
-    static std::vector<std::vector<State>> targets;
-    static std::vector<State> e_sources;
-    static std::vector<State> e_targets;
-    if (alphasize>sources.size()) {
-        sources.resize(alphasize);
-        targets.resize(alphasize);
-    }
-
-    e_sources.clear();
-    e_targets.clear();
-
-    //WHEN ONLY MAX SYMBOL IS COMPUTED
-    // for (int i = 0;i<alphasize;i++) {
-    //     for (int i = 0;i<alphasize;i++) {
-    //         if (!sources[i].empty())
-    //         {
-    //             sources[i].resize(0);
-    //             targets[i].resize(0);
-    //         }
-    //     }
-    // }
-
-    //WHEN ALL SYMBOLS ARE COMPUTED
-    for (Symbol symbol: symbols) {
-        if(!sources[symbol].empty()) {
-            sources[symbol].clear();
-            targets[symbol].clear();
-        }
-    }
-#else
-    // NORMAL, NON STATIC DATA STRUCTURES:
-    //All transition of delta are to be copied here, into two arrays of transition sources and targets indexed by the transition symbol.
-    // There is a special treatment for epsilon, since we want the arrays to be only as long as the largest symbol in the automaton,
-    // and epsilon is the maximum (so we don't want to have the maximum array lenght whenever epsilon is present)
-    std::vector<std::vector<State>> sources (alphasize);
-    std::vector<std::vector<State>> targets (alphasize);
-    std::vector<State> e_sources;
-    std::vector<State> e_targets;
-#endif
-
-    //Copy all transition with non-e symbols to the arrays of sources and targets indexed by symbols.
-    //Targets and sources of e-transitions go to the special place.
-    //Important: since we are going through delta in order of sources, the sources arrays are all ordered.
-    for (State sourceState{ 0 }; sourceState < num_of_states; ++sourceState) {
-        for (const Move &move: aut.delta[sourceState]) {
-            if (move.symbol == EPSILON) {
-                for (const State targetState: move.targets) {
-                    //reserve_on_insert(e_sources);
-                    e_sources.push_back(sourceState);
-                    //reserve_on_insert(e_targets);
-                    e_targets.push_back(targetState);
-                }
-            }
-            else {
-                for (const State targetState: move.targets) {
-                    //reserve_on_insert(sources[move.symbol]);
-                    sources[move.symbol].push_back(sourceState);
-                    //reserve_on_insert(targets[move.symbol]);
-                    targets[move.symbol].push_back(targetState);
-                }
-            }
-        }
-    }
-
-    //Now make the delta of the reversed automaton.
-    //Important: since sources are ordered, when adding them as targets, we can just push them back.
-    result.delta.reserve(num_of_states);
-
-    // adding non-e transitions
-    for (const Symbol symbol: symbols) {
-        for (size_t i{ 0 }; i < sources[symbol].size(); ++i) {
-            State tgt_state =sources[symbol][i];
-            State src_state =targets[symbol][i];
-            Post & src_post = result.delta.get_mutable_post(src_state);
-            if (src_post.empty() || src_post.back().symbol != symbol) {
-                src_post.push_back(Move(symbol));
-            }
-            src_post.back().push_back(tgt_state);
-        }
-    }
-
-    // adding e-transitions
-    for (size_t i{ 0 }; i < e_sources.size(); ++i) {
-        State tgt_state =e_sources[i];
-        State src_state =e_targets[i];
-        Post & src_post = result.delta.get_mutable_post(src_state);
-        if (src_post.empty() || src_post.back().symbol != EPSILON) {
-            src_post.push_back(Move(EPSILON));
-        }
-        src_post.back().push_back(tgt_state);
-    }
-
-    //sorting the targets
-    //Hm I don't know why I put this here, but it should not be needed ...
-    //for (State q = 0, states_num = result.delta.post_size(); q<states_num;++q) {
-    //    for (auto m = result.delta.get_mutable_post(q).begin(); m != result.delta.get_mutable_post(q).end(); ++m) {
-    //        sort_and_rmdupl(m->targets);
-    //    }
-    //}
-
-    return result;
-}
-
-Nfa Mata::Nfa::simple_revert(const Nfa& aut) {
-    Nfa result;
-    result.clear();
-
-    const size_t num_of_states{ aut.size() };
-    result.delta.increase_size(num_of_states);
-
-    for (State sourceState{ 0 }; sourceState < num_of_states; ++sourceState) {
-        for (const Move &transition: aut.delta[sourceState]) {
-            for (const State targetState: transition.targets) {
-                result.delta.add(targetState, transition.symbol, sourceState);
-            }
-        }
-    }
-
-    result.initial = aut.final;
-    result.final = aut.initial;
-
-    return result;
-}
-
-//not so great, can be removed
-Nfa Mata::Nfa::somewhat_simple_revert(const Nfa& aut) {
-    const size_t num_of_states{ aut.size() };
-
-    Nfa result(num_of_states);
-
-    result.initial = aut.final;
-    result.final = aut.initial;
-
-    for (State sourceState{ 0 }; sourceState < num_of_states; ++sourceState) {
-        for (const Move &transition: aut.delta[sourceState]) {
-            for (const State targetState: transition.targets) {
-                Post & post = result.delta.get_mutable_post(targetState);
-                //auto move = std::find(post.begin(),post.end(),Move(transition.symbol));
-                auto move = post.find(Move(transition.symbol));
-                if (move == post.end()) {
-                    //post.push_back(Move(transition.symbol,sourceState));
-                    post.insert(Move(transition.symbol,sourceState));
-                }
-                else
-                    move->push_back(sourceState);
-                    //move->insert(sourceState);
-            }
-        }
-    }
-
-    //sorting the targets
-    for (State q = 0, states_num = result.delta.num_of_states(); q < states_num; ++q) {
-        //Post & post = result.delta.get_mutable_post(q);
-        //Util::sort_and_rmdupl(post);
-        for (auto m = result.delta.get_mutable_post(q).begin(); m != result.delta.get_mutable_post(q).end(); ++m) {
-            sort_and_rmdupl(m->targets);
-        }
-    }
-
-    return result;
-}
-
-Nfa Mata::Nfa::revert(const Nfa& aut) {
-    return simple_revert(aut);
-    //return fragile_revert(aut);
-    //return somewhat_simple_revert(aut);
-}
-
-bool Mata::Nfa::is_deterministic(const Nfa& aut)
-{
-    if (aut.initial.size() != 1) { return false; }
-
-    if (aut.delta.empty()) { return true; }
-
-    const size_t aut_size = aut.size();
-    for (size_t i = 0; i < aut_size; ++i)
-    {
-        for (const auto& symStates : aut.delta[i])
-        {
-            if (symStates.size() != 1) { return false; }
-        }
-    }
-
-    return true;
-}
-bool Mata::Nfa::is_complete(const Nfa& aut, const Alphabet& alphabet)
-{
-    Util::OrdVector<Symbol> symbs_ls = alphabet.get_alphabet_symbols();
-    Util::OrdVector<Symbol> symbs(symbs_ls);
-
-    // TODO: make a general function for traversal over reachable states that can
-    // be shared by other functions?
-    std::list<State> worklist(aut.initial.begin(),
-                              aut.initial.end());
-    std::unordered_set<State> processed(aut.initial.begin(),
-                                        aut.initial.end());
-
-    while (!worklist.empty())
-    {
-        State state = *worklist.begin();
-        worklist.pop_front();
-
-        size_t n = 0;      // counter of symbols
-        if (!aut.delta.empty()) {
-            for (const auto &symb_stateset: aut.delta[state]) {
-                ++n;
-                if (!haskey(symbs, symb_stateset.symbol)) {
-                    throw std::runtime_error(std::to_string(__func__) +
-                                             ": encountered a symbol that is not in the provided alphabet");
-                }
-
-                for (const auto &tgt_state: symb_stateset.targets) {
-                    bool inserted;
-                    tie(std::ignore, inserted) = processed.insert(tgt_state);
-                    if (inserted) { worklist.push_back(tgt_state); }
-                }
-            }
-        }
-
-        if (symbs.size() != n) { return false; }
-    }
-
-    return true;
-}
-
-std::pair<Run, bool> Mata::Nfa::get_word_for_path(const Nfa& aut, const Run& run)
-{
-    if (run.path.empty())
-    {
-        return {{}, true};
-    }
-
-    Run word;
-    State cur = run.path[0];
-    for (size_t i = 1; i < run.path.size(); ++i)
-    {
-        State newSt = run.path[i];
-        bool found = false;
-
-        if (!aut.delta.empty())
-        {
-            for (const auto &symbolMap: aut.delta[cur]) {
-                for (State st: symbolMap.targets) {
-                    if (st == newSt) {
-                        word.word.push_back(symbolMap.symbol);
-                        found = true;
-                        break;
-                    }
-                }
-
-                if (found) { break; }
-            }
-        }
-
-        if (!found)
-        {
-            return {{}, false};
-        }
-
-        cur = newSt;    // update current state
-    }
-
-    return {word, true};
-}
-
-//TODO: this is not efficient
-bool Mata::Nfa::is_in_lang(const Nfa& aut, const Run& run)
-{
-    StateSet cur(aut.initial);
-
-    for (Symbol sym : run.word)
-    {
-        cur = aut.post(cur, sym);
-        if (cur.empty()) { return false; }
-    }
-
-    return aut.final.intersects_with(cur);
-}
-
-/// Checks whether the prefix of a string is in the language of an automaton
-// TODO: slow and it should share code with is_in_lang
-bool Mata::Nfa::is_prfx_in_lang(const Nfa& aut, const Run& run)
-{
-    StateSet cur =  StateSet{ aut.initial };
-
-    for (Symbol sym : run.word)
-    {
-        if (aut.final.intersects_with(cur)) { return true; }
-        cur = aut.post(cur, sym);
-        if (cur.empty()) { return false; }
-    }
-
-    return aut.final.intersects_with(cur);
-}
-
-/// serializes Nfa into a ParsedSection
-Mata::Parser::ParsedSection Mata::Nfa::serialize(
-        const Nfa&                aut,
-        const SymbolToStringMap*  symbol_map,
-        const StateToStringMap*   state_map) {
-    (void)aut; (void)symbol_map; (void)state_map;
-    assert(false && "Unimplemented.");
-    return {};
-}
-
-bool Mata::Nfa::is_lang_empty(const Nfa& aut, Run* cex)
-{ // {{{
-    std::list<State> worklist(
-            aut.initial.begin(), aut.initial.end());
-    std::unordered_set<State> processed(
-            aut.initial.begin(), aut.initial.end());
-
-    // 'paths[s] == t' denotes that state 's' was accessed from state 't',
-    // 'paths[s] == s' means that 's' is an initial state
-    std::map<State, State> paths;
-    for (State s : worklist)
-    {	// initialize
-        paths[s] = s;
-    }
-
-    while (!worklist.empty())
-    {
-        State state = worklist.front();
-        worklist.pop_front();
-
-        if (aut.final[state])
-        {
-            // TODO process the CEX
-            if (nullptr != cex)
-            {
-                cex->path.clear();
-                cex->path.push_back(state);
-                while (paths[state] != state)
-                {
-                    state = paths[state];
-                    cex->path.push_back(state);
-                }
-
-                std::reverse(cex->path.begin(), cex->path.end());
-                cex->word = get_word_for_path(aut, *cex).first.word;
-            }
-            return false;
-        }
-
-        if (aut.delta.empty())
-            continue;
-
-        for (const auto& symb_stateset : aut.delta[state])
-        {
-            for (const auto& tgt_state : symb_stateset.targets)
-            {
-                bool inserted;
-                tie(std::ignore, inserted) = processed.insert(tgt_state);
-                if (inserted)
-                {
-                    worklist.push_back(tgt_state);
-                    // also set that tgt_state was accessed from state
-                    paths[tgt_state] = state;
-                }
-                else
-                {
-                    // the invariant
-                    assert(haskey(paths, tgt_state));
-                }
-            }
-        }
-    }
-
-    return true;
-} // is_lang_empty }}}
-
-
-Nfa Mata::Nfa::Algorithms::minimize_brzozowski(const Nfa& aut) {
-    //compute the minimal deterministic automaton, Brzozovski algorithm
-    return determinize(revert(determinize(revert(aut))));
-}
-
-Nfa Mata::Nfa::minimize(
-                const Nfa& aut,
-                const StringMap& params)
-{
-	Nfa result;
-	// setting the default algorithm
-	decltype(Algorithms::minimize_brzozowski)* algo = Algorithms::minimize_brzozowski;
-	if (!haskey(params, "algorithm")) {
-		throw std::runtime_error(std::to_string(__func__) +
-			" requires setting the \"algo\" key in the \"params\" argument; "
-			"received: " + std::to_string(params));
-	}
-
-	const std::string& str_algo = params.at("algorithm");
-	if ("brzozowski" == str_algo) {  /* default */ }
-	else {
-		throw std::runtime_error(std::to_string(__func__) +
-			" received an unknown value of the \"algo\" key: " + str_algo);
-	}
-
-	return algo(aut);
-}
-
 std::string Nfa::print_to_DOT() const {
     std::stringstream output;
     print_to_DOT(output);
@@ -1470,376 +535,6 @@ StateSet Nfa::post(const StateSet& states, const Symbol& symbol) const {
     return res;
 }
 
-Post::const_iterator Nfa::Nfa::get_epsilon_transitions(const State state, const Symbol epsilon) const {
-    assert(is_state(state));
-    return get_epsilon_transitions(get_moves_from(state), epsilon);
-}
-
-Post::const_iterator Nfa::Nfa::get_epsilon_transitions(const Post& post, const Symbol epsilon) {
-    if (!post.empty()) {
-        if (epsilon == EPSILON) {
-            const auto& back = post.back();
-            if (back.symbol == epsilon) {
-                return std::prev(post.end());
-            }
-        } else {
-            return post.find(epsilon);
-        }
-    }
-
-    return post.end();
-}
-
-Nfa Mata::Nfa::uni(const Nfa &lhs, const Nfa &rhs) {
-    Nfa unionAutomaton = rhs;
-
-    StateToStateMap thisStateToUnionState;
-    const size_t size = lhs.size();
-    for (State thisState = 0; thisState < size; ++thisState) {
-        thisStateToUnionState[thisState] = unionAutomaton.add_state();
-    }
-
-    for (State thisInitialState : lhs.initial) {
-        unionAutomaton.initial.insert(thisStateToUnionState[thisInitialState]);
-    }
-
-    for (State thisFinalState : lhs.final) {
-        unionAutomaton.final.insert(thisStateToUnionState[thisFinalState]);
-    }
-
-    for (State thisState = 0; thisState < size; ++thisState) {
-        State unionState = thisStateToUnionState[thisState];
-        for (const Move &transitionFromThisState : lhs.delta[thisState]) {
-
-            Move transitionFromUnionState(transitionFromThisState.symbol, StateSet{});
-
-            for (State stateTo : transitionFromThisState.targets) {
-                transitionFromUnionState.insert(thisStateToUnionState[stateTo]);
-            }
-
-            unionAutomaton.delta.get_mutable_post(unionState).insert(transitionFromUnionState);
-        }
-    }
-
-    return unionAutomaton;
-}
-
-Simlib::Util::BinaryRelation Mata::Nfa::Algorithms::compute_relation(const Nfa& aut, const StringMap& params) {
-    if (!haskey(params, "relation")) {
-        throw std::runtime_error(std::to_string(__func__) +
-                                 " requires setting the \"relation\" key in the \"params\" argument; "
-                                 "received: " + std::to_string(params));
-    }
-    if (!haskey(params, "direction")) {
-        throw std::runtime_error(std::to_string(__func__) +
-                                 " requires setting the \"direction\" key in the \"params\" argument; "
-                                 "received: " + std::to_string(params));
-    }
-
-    const std::string& relation = params.at("relation");
-    const std::string& direction = params.at("direction");
-    if ("simulation" == relation && direction == "forward") {
-        return compute_fw_direct_simulation(aut);
-    }
-    else {
-        throw std::runtime_error(std::to_string(__func__) +
-                                 " received an unknown value of the \"relation\" key: " + relation);
-    }
-}
-
-Nfa Mata::Nfa::reduce(const Nfa &aut, bool trim_input, StateToStateMap *state_map, const StringMap& params) {
-    if (!haskey(params, "algorithm")) {
-        throw std::runtime_error(std::to_string(__func__) +
-                                 " requires setting the \"algorithm\" key in the \"params\" argument; "
-                                 "received: " + std::to_string(params));
-    }
-
-    Nfa aut_to_reduce{ aut };
-    StateToStateMap trimmed_state_map{};
-    if (trim_input) {
-        aut_to_reduce.trim(&trimmed_state_map);
-    }
-
-    Nfa result;
-    std::unordered_map<State,State> reduced_state_map;
-    const std::string& algorithm = params.at("algorithm");
-    if ("simulation" == algorithm) {
-        result = reduce_size_by_simulation(aut_to_reduce, reduced_state_map);
-    } else {
-        throw std::runtime_error(std::to_string(__func__) +
-                                 " received an unknown value of the \"algorithm\" key: " + algorithm);
-    }
-
-    if (state_map) {
-        state_map->clear();
-        if (trim_input) {
-            for (const auto& trimmed_mapping : trimmed_state_map) {
-                const auto reduced_mapping{ reduced_state_map.find(trimmed_mapping.second) };
-                if (reduced_mapping != reduced_state_map.end()) {
-                    (*state_map)[trimmed_mapping.first] = reduced_mapping->second;
-                }
-            }
-        } else { // Input has not been trimmed, the reduced state map is the actual input to result state map.
-            *state_map = reduced_state_map;
-        }
-    }
-
-    return result;
-}
-
-Nfa Mata::Nfa::determinize(
-        const Nfa&  aut,
-        std::unordered_map<StateSet, State> *subset_map)
-{
-
-    Nfa result;
-    //assuming all sets targets are non-empty
-    std::vector<std::pair<State, StateSet>> worklist;
-    bool deallocate_subset_map = false;
-    if (subset_map == nullptr) {
-        subset_map = new std::unordered_map<StateSet,State>();
-        deallocate_subset_map = true;
-    }
-
-    result.clear();
-
-    const StateSet S0 =  StateSet(aut.initial);
-    const State S0id = result.add_state();
-    result.initial.insert(S0id);
-
-    if (aut.final.intersects_with(S0)) {
-        result.final.insert(S0id);
-    }
-    worklist.emplace_back(std::make_pair(S0id, S0));
-
-    (*subset_map)[Mata::Util::OrdVector<State>(S0)] = S0id;
-
-    if (aut.delta.empty())
-        return result;
-
-    using Iterator = Mata::Util::OrdVector<Move>::const_iterator;
-    Mata::Util::SynchronizedExistentialIterator<Iterator> synchronized_iterator;
-
-    while (!worklist.empty()) {
-        const auto Spair = worklist.back();
-        worklist.pop_back();
-        const StateSet S = Spair.second;
-        const State Sid = Spair.first;
-        if (S.empty()) {
-            // This should not happen assuming all sets targets are non-empty.
-            break;
-        }
-
-        // add moves of S to the sync ex iterator
-        // TODO: shouldn't we also reset first?
-        for (State q: S) {
-            Mata::Util::push_back(synchronized_iterator, aut.delta[q]);
-        }
-
-        while (synchronized_iterator.advance()) {
-
-            // extract post from the sychronized_iterator iterator
-            std::vector<Iterator> moves = synchronized_iterator.get_current();
-            Symbol currentSymbol = (*moves.begin())->symbol;
-            StateSet T;
-            for (auto m: moves){
-                T = T.Union(m->targets);
-            }
-
-            const auto existingTitr = subset_map->find(T);
-            State Tid;
-            if (existingTitr != subset_map->end()) {
-                Tid = existingTitr->second;
-            } else {
-                Tid = result.add_state();
-                (*subset_map)[Mata::Util::OrdVector<State>(T)] = Tid;
-                if (aut.final.intersects_with(T)) {
-                    result.final.insert(Tid);
-                }
-                worklist.emplace_back(std::make_pair(Tid, T));
-            }
-            result.delta.get_mutable_post(Sid).insert(Move(currentSymbol, Tid));
-        }
-    }
-
-    if (deallocate_subset_map) { delete subset_map; }
-
-    return result;
-}
-
-// TODO this function should the same thing as the one taking IntermediateAut or be deleted
-Nfa Mata::Nfa::construct(
-        const Mata::Parser::ParsedSection&   parsec,
-        Alphabet*                            alphabet,
-        StringToStateMap*                    state_map)
-{ // {{{
-    Nfa aut;
-    assert(nullptr != alphabet);
-
-    if (parsec.type != Mata::Nfa::TYPE_NFA) {
-        throw std::runtime_error(std::string(__FUNCTION__) + ": expecting type \"" +
-                                 Mata::Nfa::TYPE_NFA + "\"");
-    }
-
-    bool remove_state_map = false;
-    if (nullptr == state_map) {
-        state_map = new StringToStateMap();
-        remove_state_map = true;
-    }
-
-    // a lambda for translating state names to identifiers
-    auto get_state_name = [&state_map, &aut](const std::string& str) {
-        if (!state_map->count(str)) {
-            State state = aut.add_state();
-            state_map->insert({str, state});
-            return state;
-        } else {
-            return (*state_map)[str];
-        }
-    };
-
-    // a lambda for cleanup
-    auto clean_up = [&]() {
-        if (remove_state_map) { delete state_map; }
-    };
-
-
-    auto it = parsec.dict.find("Initial");
-    if (parsec.dict.end() != it)
-    {
-        for (const auto& str : it->second)
-        {
-            State state = get_state_name(str);
-            aut.initial.insert(state);
-        }
-    }
-
-
-    it = parsec.dict.find("Final");
-    if (parsec.dict.end() != it)
-    {
-        for (const auto& str : it->second)
-        {
-            State state = get_state_name(str);
-            aut.final.insert(state);
-        }
-    }
-
-    for (const auto& body_line : parsec.body)
-    {
-        if (body_line.size() != 3)
-        {
-            // clean up
-            clean_up();
-
-            if (body_line.size() == 2)
-            {
-                throw std::runtime_error("Epsilon transitions not supported: " +
-                                         std::to_string(body_line));
-            }
-            else
-            {
-                throw std::runtime_error("Invalid transition: " +
-                                         std::to_string(body_line));
-            }
-        }
-
-        State src_state = get_state_name(body_line[0]);
-        Symbol symbol = alphabet->translate_symb(body_line[1]);
-        State tgt_state = get_state_name(body_line[2]);
-
-        aut.delta.add(src_state, symbol, tgt_state);
-    }
-
-    // do the dishes and take out garbage
-    clean_up();
-
-    return aut;
-} // construct }}}
-
-Nfa Mata::Nfa::construct(
-        const Mata::IntermediateAut&         inter_aut,
-        Alphabet*                            alphabet,
-        StringToStateMap*                    state_map)
-{ // {{{
-    Nfa aut;
-    assert(nullptr != alphabet);
-
-    if (!inter_aut.is_nfa()) {
-        throw std::runtime_error(std::string(__FUNCTION__) + ": expecting type \"" +
-                                 Mata::Nfa::TYPE_NFA + "\"");
-    }
-
-    StringToStateMap tmp_state_map;
-    if (nullptr == state_map) {
-        state_map = &tmp_state_map;
-    }
-
-    // a lambda for translating state names to identifiers
-    auto get_state_name = [&state_map, &aut](const std::string& str) {
-        if (!state_map->count(str)) {
-            State state = aut.add_state();
-            state_map->insert({str, state});
-            return state;
-        } else {
-            return (*state_map)[str];
-        }
-    };
-
-    for (const auto& str : inter_aut.initial_formula.collect_node_names())
-    {
-        State state = get_state_name(str);
-        aut.initial.insert(state);
-    }
-
-    for (const auto& trans : inter_aut.transitions)
-    {
-        if (trans.second.children.size() != 2)
-        {
-            if (trans.second.children.size() == 1)
-            {
-                throw std::runtime_error("Epsilon transitions not supported");
-            }
-            else
-            {
-                throw std::runtime_error("Invalid transition");
-            }
-        }
-
-        State src_state = get_state_name(trans.first.name);
-        Symbol symbol = alphabet->translate_symb(trans.second.children[0].node.name);
-        State tgt_state = get_state_name(trans.second.children[1].node.name);
-
-        aut.delta.add(src_state, symbol, tgt_state);
-    }
-
-    std::unordered_set<std::string> final_formula_nodes;
-    if (!(inter_aut.final_formula.node.is_constant())) {
-        // we do not want to parse true/false (constant) as a state so we do not collect it
-        final_formula_nodes = inter_aut.final_formula.collect_node_names();
-    }
-    // for constant true, we will pretend that final nodes are negated with empty final_formula_nodes
-    bool final_nodes_are_negated = (inter_aut.final_formula.node.is_true() || inter_aut.are_final_states_conjunction_of_negation());
-
-    if (final_nodes_are_negated) {
-        // we add all states NOT in final_formula_nodes to final states
-        for (const auto &state_name_and_id : *state_map) {
-            if (!final_formula_nodes.count(state_name_and_id.first)) {
-                aut.final.insert(state_name_and_id.second);
-            }
-        }
-    } else {
-        // we add all states in final_formula_nodes to final states
-        for (const auto& str : final_formula_nodes)
-        {
-            State state = get_state_name(str);
-            aut.final.insert(state);
-        }
-    }
-
-    return aut;
-} // construct }}}
-
 Nfa::const_iterator Nfa::const_iterator::for_begin(const Nfa* nfa)
 { // {{{
     assert(nullptr != nfa);
@@ -1939,11 +634,6 @@ bool Nfa::const_iterator::operator==(const Nfa::const_iterator& rhs) const {
     if (this->is_end && rhs.is_end) { return true; }
     if ((this->is_end && !rhs.is_end) || (!this->is_end && rhs.is_end)) { return false; }
     return ssIt == rhs.ssIt && tlIt == rhs.tlIt && trIt == rhs.trIt;
-}
-
-std::ostream& std::operator<<(std::ostream& os, const Mata::Nfa::Nfa& nfa) {
-    nfa.print_to_mata(os);
-    return os;
 }
 
 // Other versions, maybe an interesting experiment with speed of data structures.
@@ -2092,17 +782,7 @@ void Nfa::unify_final() {
     final.insert(new_final_state);
 }
 
-void Mata::Nfa::fill_alphabet(const Mata::Nfa::Nfa& nfa, OnTheFlyAlphabet& alphabet) {
-    const size_t nfa_num_of_states{ nfa.size() };
-    for (Mata::Nfa::State state{ 0 }; state < nfa_num_of_states; ++state) {
-        for (const Move& move: nfa.delta[state]) {
-            alphabet.update_next_symbol_value(move.symbol);
-            alphabet.try_add_new_symbol(std::to_string(move.symbol), move.symbol);
-        }
-    }
-}
-
-void Nfa::add_symbols_to(OnTheFlyAlphabet& target_alphabet) const {
+void Nfa::add_symbols_to(OnTheFlyAlphabet& alphabet) const {
     size_t aut_num_of_states{size() };
     for (Mata::Nfa::State state{ 0 }; state < aut_num_of_states; ++state) {
         for (const Move& move: delta[state]) {
@@ -2221,82 +901,4 @@ OrdVector<Symbol> Nfa::get_used_symbols() const {
     //        ov.push_back(i);
     //    }
     //return ov;
-
-}
-
-Mata::OnTheFlyAlphabet Mata::Nfa::create_alphabet(const ConstAutRefSequence& nfas) {
-    Mata::OnTheFlyAlphabet alphabet{};
-    for (const auto& nfa: nfas) {
-        fill_alphabet(nfa, alphabet);
-    }
-    return alphabet;
-}
-
-Mata::OnTheFlyAlphabet Mata::Nfa::create_alphabet(const AutRefSequence& nfas) {
-    Mata::OnTheFlyAlphabet alphabet{};
-    for (const auto& nfa: nfas) {
-        fill_alphabet(nfa, alphabet);
-    }
-    return alphabet;
-}
-
-Mata::OnTheFlyAlphabet Mata::Nfa::create_alphabet(const ConstAutPtrSequence& nfas) {
-    Mata::OnTheFlyAlphabet alphabet{};
-    for (const Mata::Nfa::Nfa* const nfa: nfas) {
-        fill_alphabet(*nfa, alphabet);
-    }
-    return alphabet;
-}
-
-Mata::OnTheFlyAlphabet Mata::Nfa::create_alphabet(const AutPtrSequence& nfas) {
-    Mata::OnTheFlyAlphabet alphabet{};
-    for (const Mata::Nfa::Nfa* const nfa: nfas) {
-        fill_alphabet(*nfa, alphabet);
-    }
-    return alphabet;
-}
-
-Nfa Mata::Nfa::create_empty_string_nfa() {
-    return Nfa{ 1, StateSet{ 0 }, StateSet{ 0 } };
-}
-
-Nfa Mata::Nfa::create_sigma_star_nfa(Mata::Alphabet* alphabet) {
-    Nfa nfa{ 1, StateSet{ 0 }, StateSet{ 0 }, alphabet };
-    for (const Mata::Symbol& symbol : alphabet->get_alphabet_symbols()) {
-        nfa.delta.add(0, symbol, 0);
-    }
-    return nfa;
-}
-
-Run Mata::Nfa::encode_word(const Mata::StringToSymbolMap& symbol_map, const std::vector<std::string>& input) {
-    Run result;
-    for (const auto& str : input) { result.word.push_back(symbol_map.at(str)); }
-    return result;
-}
-
-Move& Move::operator=(Move&& rhs) noexcept {
-    if (*this != rhs) {
-        symbol = rhs.symbol;
-        targets = std::move(rhs.targets);
-    }
-    return *this;
-}
-
-void Move::insert(State s) {
-    if(targets.empty() || targets.back() < s) {
-        targets.push_back(s);
-        return;
-    }
-    // Find the place where to put the element (if not present).
-    // insert to OrdVector without the searching of a proper position inside insert(const Key&x).
-    auto it = std::lower_bound(targets.begin(), targets.end(), s);
-    if (it == targets.end() || *it != s) {
-        targets.insert(it, s);
-    }
-}
-
-void Move::insert(const StateSet& states) {
-    for (State s : states) {
-        insert(s);
-    }
 }
